DMCA Removal - read below for more details.
We are removing our Digital Millennium Copyright Act (DMCA) notices from our repositories. We have decided to take this step in order to foster a more positive and supportive environment for the open source community.

Please be assured that we will continue to monitor our resources and take action against any individuals or organizations that violate the license under which our work is released. Our DMCA notices are only used in cases where our work is being used in a manner that violates the terms of the license, and we would never send a DMCA notice to someone who is using our resources properly and in accordance with the terms of the license.

In addition, our use of DMCA notices has always been to protect the use of our trademarked name, Project Sloth, and has never been used to send illegitimate takedowns. Rest assured that we will continue to protect our projects with DMCA notices on the backend to ensure that they are not used in violation of the license.

We appreciate your support and interest in our work, and we are always happy to answer any questions you may have about how to use our resources properly.

The Project Sloth Team